+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/settings/RoomDevInfoViewModel.kt
@@ -1,8 +1,15 @@
 package de.connect2x.trixnity.messenger.viewmodel.room.settings
 
+import de.connect2x.trixnity.client.room
+import de.connect2x.trixnity.client.store.OlmCryptoStore
 import de.connect2x.trixnity.core.model.RoomId
 import de.connect2x.trixnity.messenger.util.BackCallback
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.launch
 
 interface RoomDevInfoViewModelFactory {
     fun create(
@@ -18,6 +25,9 @@ interface RoomDevInfoViewModelFactory {
 
 interface RoomDevInfoViewModel {
     val roomId: RoomId
+    val showResetEncryptionButton: StateFlow<Boolean>
+
+    fun resetEncryption()
 
     fun back()
 }
@@ -30,8 +40,24 @@ class RoomDevInfoViewModelImpl(
 
     private val backCallback = BackCallback { onBack() }
 
+    private val olmCryptoStore: OlmCryptoStore = matrixClient.di.get<OlmCryptoStore>()
+
+    override val showResetEncryptionButton =
+        matrixClient.room
+            .getById(roomId)
+            .map { it?.encrypted == true }
+            .stateIn(coroutineScope, SharingStarted.Eagerly, false)
+
     init {
         registerBackCallback(backCallback)
+    }
+
+    override fun resetEncryption() {
+        coroutineScope.launch {
+            if (showResetEncryptionButton.value) {
+                olmCryptoStore.updateOutboundMegolmSession(roomId) { null }
+            }
+        }
     }
 
     override fun back() {