+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- UI: Ability to reset encryption keys for a room in room developer settings
+
 ### Changed
 
 ### Deprecated